limpando o código do pacBase
