+++ exibMapa.py
@@ -1,0 +1,25 @@
+import numpy as np
+import os
+from cont_pontuacao import exib_pontos
+mapa = np.array([
+    ["▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉"],
+    ["▉"," "," "," "," ","▉"," "," "," "," "," "," "," "," "," "," ","▉"," "," "," "," ","▉"],
+    ["▉"," ","▉","▉"," ","▉"," ","▉","▉","▉","▉","▉","▉","▉","▉"," ","▉"," ","▉","▉"," ","▉"],
+    ["▉"," ","▉"," "," "," "," "," "," "," "," "," "," "," "," "," "," "," "," ","▉"," ","▉"],
+    ["▉"," ","▉"," ","▉","▉"," ","▉"," "," "," "," "," "," ","▉"," ","▉","▉"," ","▉"," ","▉"],
+    ["▉"," ","▉"," "," "," "," ","▉"," "," ","①","②","③"," ","▉"," "," "," "," ","▉"," ","▉"],
+    ["▉"," "," "," ","▉","▉"," ","▉","▉","▉","▉","▉","▉","▉","▉"," ","▉","▉"," "," "," ","▉"],
+    ["▉"," ","▉"," "," "," "," "," "," "," "," "," "," "," "," "," "," "," "," ","▉"," ","▉"],
+    ["▉"," ","▉"," ","▉","▉"," ","▉","▉","▉","▉","▉","▉","▉","▉"," ","▉","▉"," ","▉"," ","▉"],
+    ["▉"," ","▉"," "," "," "," "," "," "," "," "," "," "," "," "," "," "," "," ","▉"," ","▉"],
+    ["▉"," ","▉","▉"," ","▉"," ","▉","▉","▉","▉","▉","▉","▉","▉"," ","▉"," ","▉","▉"," ","▉"],
+    ["▉"," "," "," "," ","▉"," "," "," "," "," "," "," "," "," "," ","▉"," "," "," ","c","▉"],
+    ["▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉","▉"]
+])
+
+# Função para exibir o mapa
+def exibir_mapa():
+    os.system('cls' if os.name == 'nt' else 'clear')
+    for linha in mapa:
+        print("".join(linha))
+    exib_pontos()
