--- txtsinistro.txt
@@ -1,1 +1,0 @@
-Implementar ação
